refactor(): improve naming and directories structure
