+++ lib/simplecov_linter_formatter/exporters/json_result_exporter.rb
