+++ lib/simplecov_linter_formatter/exporters/text_result_exporter.rb
