+++ spec/lib/exporters/json_result_exporter_spec.rb
