+++ spec/lib/exporters/result_exporter_spec.rb
